tests: UserController

+++ src/main/java/dev/feramaro/simplified_picpay/services/AuthorizationService.java
@@ -2,7 +2,9 @@ package dev.feramaro.simplified_picpay.services;
 
 import dev.feramaro.simplified_picpay.domain.user.User;
 
+import java.math.BigDecimal;
+
 public interface AuthorizationService {
 
-    boolean authorize(User user, Double amount);
+    boolean authorize(User user, BigDecimal amount);
 }
